agnes/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
openai>=1.14.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
tavily-python>=0.3.0
scikit-learn>=1.4.0
numpy>=1.26.0
python-dotenv>=1.0.0
